Moved include folder to the root
