+++ scripts/setup_static_mac_sdl.sh
@@ -2,7 +2,7 @@
 VERSION=2.0.12
 FOLDER=special
 DOME_DIR=$PWD
-INCLUDE_DIR=$DOME_DIR/src/include
+INCLUDE_DIR=$DOME_DIR/include
 LIB_DIR=$DOME_DIR/lib
 DIRECTORY=$LIB_DIR/SDL2-${VERSION}
